Backend/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7

# CORS support
starlette>=0.26.1

# Image processing
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.2

# YOLOv5 requirements
torch>=1.8.0
torchvision>=0.9.0

# If you're using CUDA, you might want to specify:
# torch>=1.8.0+cu111
# torchvision>=0.9.0+cu111

# Optional: for deployment
python-multipart>=0.0.6
